Client: Reorganize services.
