+++ client/src/app/core/services/communication/webrtc.service.spec.ts
@@ -1,5 +1,5 @@
 import { TestBed } from '@angular/core/testing';
-import { TestImports, TestProviders } from '../../testing/test-helper';
+import { TestImports, TestProviders } from '../../../testing/test-helper';
 
 describe('WebRTCService', () => {
   beforeEach(() => {
