+++ client/src/app/core/services/communication/websocket-connection.service.spec.ts
@@ -1,11 +1,12 @@
 import { TestBed } from '@angular/core/testing';
-import { TestImports, TestProviders } from '../../testing/test-helper';
+import { provideRouter } from '@angular/router';
+import { TestImports, TestProviders } from '../../../testing/test-helper';
 
-describe('ChatService', () => {
+describe('WebSocketConnectionService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [...TestImports],
-      providers: [...TestProviders],
+      providers: [...TestProviders, provideRouter([])],
     });
   });
