+++ client/src/app/core/services/migration/migration.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
-import { TestImports, TestProviders } from '../../testing/test-helper';
+import { TestImports, TestProviders } from '../../../testing/test-helper';
 
-describe('RoomService', () => {
+describe('MigrationService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [...TestImports],
