+++ client/src/app/features/chat/chat.component.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { TestImports, TestProviders } from '../../testing/test-helper';
 import { PLATFORM_ID } from '@angular/core';
-import { FlowbiteService } from '../../core/services/flowbite.service';
+import { FlowbiteService } from '../../core/services/ui/flowbite.service';
 import { ActivatedRoute, convertToParamMap, provideRouter } from '@angular/router';
 import { of } from 'rxjs';
